// frontend/NourishNet/Models/Enums/DistrictValue.ts
export enum District {
    AMPARA = 'AMPARA', 
    ANURADHAPURA = 'ANURADHAPURA', 
    BADULAA = 'BADULAA', 
    BATTICALOA = 'BATTICALOA', 
    COLOMBO = 'COLOMBO',
    GALLE = 'GALLE', 
    GAMPAHA = 'GAMPAHA', 
    HAMBANTHOTA = 'HAMBANTHOTA', 
    JAFFNA = 'JAFFNA', 
    KALUTHARA = 'KALUTHARA',
    KANDY = 'KANDY', 
    KEGALLE = 'KEGALLE', 
    KILINOCHCHI = 'KILINOCHCHI', 
    KURUNEGALA = 'KURUNEGALA', 
    MANNER = 'MANNER',
    MATALE = 'MATALE', 
    MATARA = 'MATARA', 
    MONARAGALA = 'MONARAGALA', 
    MULLAITIVU = 'MULLAITIVU', 
    NUWARA_ELIYA = 'NUWARA_ELIYA',
    POLLANNARUWA = 'POLLANNARUWA', 
    PUTTALAM = 'PUTTALAM', 
    RATHNAPURA = 'RATHNAPURA', 
    TRINCOMALEE = 'TRINCOMALEE', 
    VAVUNIYA = 'VAVUNIYA'
}